Extract bytecode-IR->Go translator into importable codegen package
cmd/kl's bc->go native now wraps codegen.CodeGenerator so other build
tools (the upcoming cmd/yggdrasil-build) can drive the translator
in-process without going through KL. No behavior change.
